% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helloworld.R
\name{helloworld}
\alias{helloworld}
\title{This is the function for showing the information of XXX}
\usage{
helloworld()
}
\arguments{
\item{details}{A character string to convert.}
}
\value{
A character string with emojis.
}
\description{
This is the function for showing the information of XXX
}
\examples{
hello()

}
